+++ src/stacks/api-gateway/ApiGatewayStack.ts
@@ -1,14 +1,7 @@
 import path from 'path';
 import { Stack, StackProps } from 'aws-cdk-lib';
-import {
-  CorsHttpMethod,
-  HttpApi,
-  HttpMethod,
-} from 'aws-cdk-lib/aws-apigatewayv2';
-import {
-  HttpLambdaAuthorizer,
-  HttpLambdaResponseType,
-} from 'aws-cdk-lib/aws-apigatewayv2-authorizers';
+import { CorsHttpMethod, HttpApi, HttpMethod } from 'aws-cdk-lib/aws-apigatewayv2';
+import { HttpLambdaAuthorizer, HttpLambdaResponseType } from 'aws-cdk-lib/aws-apigatewayv2-authorizers';
 import { HttpLambdaIntegration } from 'aws-cdk-lib/aws-apigatewayv2-integrations';
 import { UserPool } from 'aws-cdk-lib/aws-cognito';
 import { Table } from 'aws-cdk-lib/aws-dynamodb';
@@ -18,7 +11,7 @@ import { RetentionDays } from 'aws-cdk-lib/aws-logs';
 import { StringParameter } from 'aws-cdk-lib/aws-ssm';
 import { Construct } from 'constructs';
 import { TABLES } from '../core/DynamoDBStack';
-import { REPORT_TABLE_NAME } from './api/express-app/interfaces';
+import { REPORT_TABLE_NAME } from './api-lambda/express-app/interfaces';
 
 interface ApiGatewayStackProps extends StackProps {
   userPool: UserPool;
@@ -30,35 +23,20 @@ export class ApiGatewayStack extends Stack {
   constructor(scope: Construct, id: string, props: ApiGatewayStackProps) {
     super(scope, id, props);
 
-    const reportTable = Table.fromTableName(
-      this,
-      'reportTable',
-      StringParameter.valueForStringParameter(
-        this,
-        TABLES.REPORT_TABLE_PARAMETER_NAME,
-      ),
-    );
+    const reportTable = Table.fromTableName(this, 'reportTable', StringParameter.valueForStringParameter(this, TABLES.REPORT_TABLE_PARAMETER_NAME));
 
-    const httpLambdaAuthorizer = new NodejsFunction(
-      this,
-      'httpLambdaAuthorizer',
-      {
-        runtime: Runtime.NODEJS_20_X,
-        logRetention: RetentionDays.ONE_MONTH,
-      },
-    );
-
-    const authorizer = new HttpLambdaAuthorizer(
-      'authorizer',
-      httpLambdaAuthorizer,
-      {
-        responseTypes: [HttpLambdaResponseType.SIMPLE],
-      },
-    );
-
-    const report = new NodejsFunction(this, 'report', {
+    const httpLambdaAuthorizer = new NodejsFunction(this, 'httpLambdaAuthorizer', {
       runtime: Runtime.NODEJS_20_X,
-      entry: path.join(__dirname, './report/serverless.ts'),
+      logRetention: RetentionDays.ONE_MONTH,
+    });
+
+    const authorizer = new HttpLambdaAuthorizer('authorizer', httpLambdaAuthorizer, {
+      responseTypes: [HttpLambdaResponseType.SIMPLE],
+    });
+
+    const apiLambda = new NodejsFunction(this, 'apiLambda', {
+      runtime: Runtime.NODEJS_20_X,
+      entry: path.join(__dirname, './api-lambda/serverless.ts'),
       logRetention: RetentionDays.ONE_MONTH,
       environment: { [REPORT_TABLE_NAME]: reportTable.tableName },
       // timeout: Duration.seconds(15),
@@ -76,10 +54,10 @@ export class ApiGatewayStack extends Stack {
     this.api.addRoutes({
       path: '/api/{proxy+}',
       methods: [HttpMethod.ANY],
-      integration: new HttpLambdaIntegration('integration', report),
+      integration: new HttpLambdaIntegration('integration', apiLambda),
       authorizer: authorizer,
     });
 
-    reportTable.grantReadWriteData(report);
+    reportTable.grantReadWriteData(apiLambda);
   }
 }
