+++ src/stacks/api-gateway/api-lambda/debug/report-api.http
