+++ src/stacks/api-gateway/api-lambda/express-app/config/dynamoDB.ts
