+++ src/stacks/api-gateway/api-lambda/express-app/interfaces.ts
